finished frontend backend communication

+++ src/slogo/UI/Turtle.java
@@ -14,7 +14,7 @@ public class Turtle {
 	private Node turtleImage;
 	private double xpos;
 	private double ypos;
-	private int orientation;
+	private double orientation;
 	private Map<String, Node> imagesMap;
 
 	public Turtle(String s, int x, int y){
@@ -76,7 +76,7 @@ public class Turtle {
 		return ypos;
 	}
 
-	protected void setOrientation(int o){
+	protected void setOrientation(double o){
 		turtleImage.setRotate(o);
 		orientation = o;
 	}
@@ -86,7 +86,7 @@ public class Turtle {
 		orientation = orientation + o;
 	}
 
-	protected int getOrientation(){
+	protected double getOrientation(){
 		return orientation;
 	}
